Use old generators for test mod
